Support aws vmdk
The messy aspect here is the two levels of compression:
`.xz` is compressed, but then the `streamOptimized` option
for the vmdk format includes internal compression.

Reproducing the inner vmdk compression is hard.  What I think
will help is for us to add a new `qcow2-uncompressed-checksum`
to the stream metadata.

But even then, we're now relying on `qemu-img` existing in
the target container (which may be different from the version
used to convert the image originally).